moving logic to a composable
